pyproject.toml
[project]
name = "claude-mcp-bridge"
version = "0.1.0"
description = "MCP server bridging claude.ai to Claude CLI on a remote server"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "mcp>=1.26",
    "uvicorn>=0.32",
    "structlog>=24.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "anyio>=4.0",
]

[project.scripts]
mcp-bridge = "mcp_bridge.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_bridge"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
